jQuery updates in preferences & auto IP detection
See readme.
That commit correspond to v0.2b AND v0.3.
Is added: the auto-detection of ipaddress thanks to arp command and the
MAC Address.
Some jQuery elements in the preference screen (almost finish... will be
OK for v0.2c/v0.4).

Why 2 version at the same time? Because I’m hesitating on doing some
branches of the project: one focused on JS validation/ UI and one
focused on functionality.